refactor: 디자인 컬러 변경 및 코드 리팩토링
