+++ src/components/Bar/Topbar.module.scss
@@ -1,36 +1,30 @@
 @import '@styles/constants';
 
 .root {
-  position: fixed;
-  top: 0;
-  left: 0;
-
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-
-  width: 100%;
-  height: 3rem;
-  box-shadow: 0px 8px 8px -8px rgba(0, 0, 0, 0.15);
-}
-
-.title {
-  font-size: 1rem;
-  font-weight: bold;
-  color: $primary;
-  background-color: $main__white;
-}
-
-.category {
+  position: relative;
   display: flex;
   justify-content: space-between;
   align-items: center;
+  padding: 0 18px;
 
-  width: 30%;
-  margin: 0 1rem;
+  height: 48px;
+  box-shadow: 0px 4px 6px rgba(0, 0, 0, 0.05), 0px 1px 3px rgba(0, 0, 0, 0.1),
+    inset 0px 1px 0px rgba(255, 255, 255, 0.1);
+}
+
+.logo {
+  font-size: 18px;
+  font-weight: 700;
+  color: $primary;
+}
+
+.category {
+  flex-shrink: 0;
 
   & button {
-    background-color: $main__white;
+    font-size: 14px;
+    padding: 0 24px;
+    color: $secondary;
 
     &:hover {
       color: $primary;
@@ -38,10 +32,21 @@
   }
 }
 
-.my {
-  background-color: $main__white;
+.user {
+  flex-shrink: 0;
+  height: 36px;
+  border-radius: 18px;
+  border: 1px $secondary solid;
+  padding: 0 12px;
 
-  &:hover {
+  & button {
+    font-size: 12px;
+    line-height: 3;
     color: $primary;
+  }
+
+  & button:first-child:after {
+    content: '|';
+    margin: 0 8px;
   }
 }
