+++ src/components/Form/SigninForm.jsx
@@ -19,7 +19,7 @@ const SigninForm = (props) => {
 
   return (
     <form className={className} onSubmit={onSubmit}>
-      <h1 className={styles.title}>Login</h1>
+      <h1 className={styles.title}>로그인</h1>
       <input
         className={styles.input__large}
         type="text"
