+++ src/components/Form/index.js
@@ -1,2 +1,3 @@
 export { default as SigninForm } from './SigninForm';
 export { default as SignupForm } from './SignupForm';
+export { default as MyInfoEditForm } from './MyInfoEditForm';
