--- src/components/Layout/Layout.module.scss
@@ -1,9 +1,0 @@
-.root {
-  width: 100%;
-  height: calc(100vh - 3rem);
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  margin-top: 3rem;
-}
