+++ src/components/Layout/LayoutColumn.module.scss
@@ -1,8 +1,8 @@
 .root {
   width: 100%;
-  height: calc(100vh - 3rem);
+  height: calc(100vh - 48px);
   display: flex;
-  justify-content: flex-start;
-  align-items: flex-start;
-  margin-top: 3rem;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
 }
