+++ src/components/Layout/LayoutRow.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
-import styles from './Layout.module.scss';
+import styles from './LayoutRow.module.scss';
 import classNames from 'classnames/bind';
 
 const cx = classNames.bind(styles);
 
-const Layout = (props) => {
+const LayoutRow = (props) => {
   const { children, className: rootClassName } = props;
 
   const className = cx(styles.root, rootClassName);
@@ -12,4 +12,4 @@ const Layout = (props) => {
   return <div className={className}>{children}</div>;
 };
 
-export default Layout;
+export default LayoutRow;
