+++ src/components/Layout/LayoutRow.module.scss
@@ -1,0 +1,6 @@
+.root {
+  height: calc(100vh - 48px);
+  display: flex;
+  justify-content: flex-start;
+  align-items: flex-start;
+}
