+++ src/components/Layout/index.js
@@ -1,2 +1,2 @@
-export { default as Layout } from './Layout';
 export { default as LayoutColumn } from './LayoutColumn';
+export { default as LayoutRow } from './LayoutRow';
