--- src/components/User/index.js
@@ -1,1 +1,0 @@
-export { default as MyInfoEdit } from './MyInfoEdit';
