+++ src/containers/Form/MyInfoEditForm.container.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { MyInfoEdit } from '@components/User';
+import { MyInfoEditForm } from '@components/Form';
 import { useForm } from '@hooks';
 
 const MyInfoEditContainer = (props) => {
@@ -10,7 +10,7 @@ const MyInfoEditContainer = (props) => {
       validate: () => {},
     });
   return (
-    <MyInfoEdit
+    <MyInfoEditForm
       values={values}
       setValues={setValues}
       onChange={handleChange}
