+++ src/containers/Form/index.js
@@ -1,2 +1,3 @@
 export { default as SigninFormContainer } from './SigninForm.container';
 export { default as SignupFormContainer } from './SignupForm.container';
+export { default as MyInfoEditFormContainer } from './MyInfoEditForm.container';
