--- src/containers/User/index.js
@@ -1,1 +1,0 @@
-export { default as MyInfoEditContainer } from './MyInfoEdit.container';
