+++ src/pages/SigninPage.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import { SigninFormContainer } from '@containers/Form';
-import { Layout } from '@components/Layout';
+import { LayoutColumn } from '@components/Layout';
 
 const SigninPage = (props) => {
   return (
-    <Layout>
+    <LayoutColumn>
       <SigninFormContainer />
-    </Layout>
+    </LayoutColumn>
   );
 };
