+++ src/styles/constants.scss
@@ -1,4 +1,4 @@
-$primary: #0099cc;
+$primary: #076486;
 $secondary: #8c8d91;
 $point: #5abfbc;
 $main__gray: #c8c8c8;
